work in progress support for fluent
