+++ Cargo.toml
@@ -8,7 +8,7 @@ license = "MIT"
 name = "cargo-i18n"
 readme = "README.md"
 repository = "https://github.com/kellpossible/cargo-i18n"
-version = "0.1.10"
+version = "0.2.0"
 
 [badges]
 maintenance = { status = "actively-developed" }
@@ -16,9 +16,9 @@ maintenance = { status = "actively-developed" }
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-i18n-embed = { version = "0.6", path = "./i18n-embed", features = ["desktop-requester"] }
-i18n-build = { version = "0.4", path = "./i18n-build", features = ["localize"] }
-i18n-config = { version = "0.2", path = "./i18n-config" }
+i18n-embed = { version = "0.7", path = "./i18n-embed", features = ["desktop-requester"] }
+i18n-build = { version = "0.5", path = "./i18n-build", features = ["localize"] }
+i18n-config = { version = "0.3", path = "./i18n-config" }
 anyhow = "1.0"
 gettext = "0.4"
 tr = { version = "0.1", default-features = false, features = ["gettext"] }
