+++ README.md
@@ -149,12 +149,10 @@ For a complete example usage, including localizing sub-crates as libraries, you 
 Available configuration options for `i18n.toml`:
 
 ```toml
-# (Required) The locale/language identifier of the language used in the source
-# code.
-src_locale = "en-US"
-
-# (Required) The locales that the software will be translated into.
-target_locales = ["es", "ru", "cz"]
+# (Required) The locale identifier of the language used in the source
+# code for gettext system, and the primary fallback locale (for which
+# all strings must be present) when using the fluent system.
+fallback_locale = "en-US"
 
 # (Optional) Specify which subcrates to perform localization within. If the
 # subcrate has its own `i18n.toml` then, it will have its localization
@@ -164,6 +162,9 @@ subcrates = ["subcrate1", "subcrate2"]
 
 # (Optional) Use the gettext localization system.
 [gettext]
+# (Required) The locales that the software will be translated into.
+target_locales = ["es", "ru", "cz"]
+
 # (Required) Path to the output directory, relative to `i18n.toml` of the crate
 # being localized.
 output_dir = "i18n"
