+++ i18n-build/i18n/po/ru/i18n_build.po
@@ -2,7 +2,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: i18n-build\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2020-04-17 08:38+0000\n"
+"POT-Creation-Date: 2020-08-20 07:45+0000\n"
 "Language: ru\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
@@ -51,11 +51,11 @@ msgstr "Не удалось переименовать {0} \"{1}\" в \"{2}\" т
 msgid "The path \"{0}\" is not inside the \"{1}\" directory: \"{2}\"."
 msgstr "Путь \"{0}\" не находится \"{1}\" в каталоге: \"{2}\"."
 
-#: i18n-build/src/gettext_impl/mod.rs:215
+#: i18n-build/src/gettext_impl/mod.rs:208
 msgid "There was a problem executing the \"{0}\" command"
 msgstr "Произошла ошибка при выполнении команды \"{0}\"."
 
-#: i18n-build/src/gettext_impl/mod.rs:442
+#: i18n-build/src/gettext_impl/mod.rs:443
 msgid "There was a problem parsing one of the subcrates: \"{0}\"."
 msgstr "При анализе одного из подкрэйтора произошла ошибка: \"{0}\"."
