+++ i18n-config/Cargo.toml
@@ -8,7 +8,7 @@ license = "MIT"
 name = "i18n-config"
 readme = "README.md"
 repository = "https://github.com/kellpossible/cargo-i18n/tree/master/i18n-config"
-version = "0.2.2"
+version = "0.3.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
