+++ i18n-config/src/fluent.rs
@@ -1,0 +1,8 @@
+use serde::Deserialize;
+
+/// The data structure representing what is stored (and possible to
+/// store) within the `gettext` subsection of a `i18n.toml` file.
+#[derive(Deserialize, Debug, Clone)]
+pub struct FluentConfig {
+
+}
