+++ i18n-embed/i18n-embed-impl/Cargo.toml
@@ -7,7 +7,7 @@ license = "MIT"
 name = "i18n-embed-impl"
 readme = "../README.md"
 repository = "https://github.com/kellpossible/cargo-i18n/tree/master/i18n-embed"
-version = "0.4.0"
+version = "0.5.0"
 
 [badges]
 maintenance = { status = "actively-developed" }
@@ -17,10 +17,10 @@ maintenance = { status = "actively-developed" }
 proc-macro = true
 
 [dependencies]
-quote = "1.0.3"
+quote = "1.0"
 serde = { version = "1", features = ["derive"] }
 serde_derive = "1"
-i18n-config = { version = "0.2", path = "../../i18n-config" }
+i18n-config = { version = "0.3", path = "../../i18n-config" }
 proc-macro2 = "1.0"
 
 [dependencies.syn]
