+++ i18n-embed/src/util.rs
@@ -1,0 +1,5 @@
+/// Get the translation domain from the module path (first module in
+/// the module path).
+pub fn domain_from_module(module_path: &str) -> &str {
+    module_path.split("::").next().unwrap_or(module_path)
+}
